Use `size_t` for `CImageInfo::m_Width` and `m_Height`

+++ src/engine/graphics.h
@@ -78,12 +78,12 @@ public:
 	/**
 	 * Contains the width of the image
 	 */
-	int m_Width = 0;
+	size_t m_Width = 0;
 
 	/**
 	 * Contains the height of the image
 	 */
-	int m_Height = 0;
+	size_t m_Height = 0;
 
 	/**
 	 * Contains the format of the image.
@@ -120,7 +120,7 @@ public:
 
 	size_t DataSize() const
 	{
-		return (size_t)m_Width * m_Height * PixelSize(m_Format);
+		return m_Width * m_Height * PixelSize(m_Format);
 	}
 
 	static EImageFormat ImageFormatFromInt(int Format)
